in ChapterController add delete_checked method and filter_chapters , in GradeController edit destroy method check if grade have chapters or not before delete this grade , in chapters view add button to delete checked chapters - and create form for this delete - add form for filtering , add two routes delete_checked_chapters and filter_chapters ,

+++ app/Http/Controllers/Grades/GradeController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Grades;
 
 use App\Http\Controllers\Controller;
+use App\Models\Chapter;
 use App\Models\Grade;
 use Illuminate\Http\Request;
 
@@ -138,8 +139,14 @@ class GradeController extends controller
     public function destroy(Request $request)
     {
         $id = $request->id;
-        Grade::find($id)->delete();
-        toastr()->error(trans('messages.delete'));
-        return back();
+        $chapter_id = Chapter::where('grade_id',$id)->pluck('grade_id');
+        if ($chapter_id->count() > 0){
+            toastr()->warning('لا يمكن اكمال عمليه الحذف بسبب ارتباط المرحله بصفوف دراسيه يرجي حذفها اولا !');
+            return back();
+        }else{
+            Grade::find($id)->delete();
+            toastr()->error(trans('messages.delete'));
+            return back();
+        }
     }
 }
